Rewrite image fetching and caching
Rewrite of artist and album images fetching and caching functionality.
The main objectives of this rewrite are to:
reduce images flickering when scrolling,
ensure that correct images are displayed next to corresponding artist
and to simplify the code so that it’s easier to extend.
The image downloading logic, which was previously scattered throughout
the application, got replaced with the call to ImageUtils.setArtistImage()
static function which delegates the the task to the ImageProvider singleton.
To prevent image flickering bitmaps cached in memory are applied
to ImageViews synchronously on the UI thread.
Images requiring disc or network access are processed in AsyncTasks.

Conflicts:
	src/com/andrew/apollo/service/ApolloService.java

Change-Id: Icba3939b87df72986f56e2305cbc59df9feb20f9